Remove the voice player event operation legacy
