+++ src/hooks/VoicePlayer/useVoicePlayer.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
-import { useVoicePlayerContext, VoicePlayerStatus } from ".";
+import { useVoicePlayerContext } from ".";
 
 import { VOICE_MESSAGE_FILE_NAME, VOICE_MESSAGE_MIME_TYPE } from "../../utils/consts";
-import { AudioStorageUnit, AudioUnitDefaultValue } from "./dux/initialState";
+import { AudioStorageUnit, AudioUnitDefaultValue, VoicePlayerStatus } from "./dux/initialState";
 import { generateGroupKey } from "./utils";
 
 export interface UseVoicePlayerProps {
